python/generate_custom_jwt_token/requirements.txt
boto3==1.38.23
PyJWT==2.10.1
cryptography==45.0.3